# Fixed a bug that added empty elements & changed to serial processing